+++ ZeroLibraries/Common/Math/BoolVector4.cpp
@@ -27,13 +27,13 @@ BoolVector4::BoolVector4(bool x_, bool y_, bool z_, bool w_)
   w = w_;
 }
 
-bool& BoolVector4::operator[](uint index)
+bool& BoolVector4::operator[](size_t index)
 {
   ErrorIf(index > 2, "Math::BoolVector4 - Subscript out of range.");
   return array[index];
 }
 
-bool BoolVector4::operator[](uint index) const
+bool BoolVector4::operator[](size_t index) const
 {
   ErrorIf(index > 2, "Math::BoolVector4 - Subscript out of range.");
   return array[index];
